Add support for whitelist (resolves #10, #53) (#58)
* Use session cookie instead of database

* Get rid of more old db stuff

* Get refresh token properly; Add some debugging logging

* Set token attribute properly, add more debugging

* Re order saving tokens

* Make cookies secure

* Encrypt refresh token; also add some debugging stuff

* Store access_token, not all tokens

also remove some debug stuff

* Add whitelist suport and some debug to comfirm

* Update version of python2.7

Boto was crashing because a dependency required a slightly less
antiquated version of python. Boto is used by bouncer

* Try adding a newer urllib3 dep

Hopefully this will resolve the weird boto error

* Document why urllib3 is now a faux dependency

* Remove outdated comment in dockerfile

* Clean up fixme comments and print statements

* Use User() instead of current_user which defaults

to an anonymous user during the check_session call

* Remove unused else

* Examine the cookie passed to check_session

Before, I was expecting the session variable to be valid but it's not
because of the weird way this endpoint is hit. We need to check the
cookie in the state that it's passed in, and this means decoding it.

* Test encryption and email redaction

* Add testing instructions to README

* Fix clean-up bug and unused requirements

* Lock down dependencies

* Remove more old requirements