Upgrade nginx version (#69)
After changing the version, Postgres failed to install during the build
of the Docker image. The error made me realize we don't need Postgres,
so I removed that as well. These changes are all in Dockerfile.

Then to be safe, removed a bunch of Python/HTML code that is no longer
being used, and that has had dependencies on Postgres, e.g., invoicing
service that stored/read data.