Access token not getting refreshed properly (fixes #75) (#78)
When UI makes API call, it calls azul-webservice with a cookie, which
in turn calls `/authorization` endpoint in app.py. Because there was
a cookie, Authorization header is not sent, and hence `token` is `None`
when `get_user_info` is invoked. The code was immediately raising an
exception when the token was `None`; instead it should raise the exception
if the token is not `None`