User must log in with Google to view data (resolves #86)
In order to contain accidental or malicious abuse of the export
function in a `dev` instance of the Commons Boardwalk Data Browser
we now force the user to log in with Google in order to view any
data, and to use the export function. If a user isn't logged in
no data can be viewed in the Data Browser.

That way a user can be tracked in case of need.